Create command to withdraw money
This command will be executed by wallet aggregate, if wallet has
founds, withdraw will remove some amount from wallet.

+++ lib/bank/accounts/projections/wallet.ex
@@ -7,7 +7,7 @@ defmodule Bank.Accounts.Projections.Wallet do
   schema "wallets" do
     field :user_uuid, :binary_id
     field :username, :string
-    field :balance, :decimal
+    field :balance, :float
 
     timestamps()
   end
